pyproject.toml
[project]
name = "janitor"
version = "0.1.0"
description = "Proxmox AI Monitoring Agent — monitors servers, debugs issues, and acts through Discord"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "proxmoxer>=2.0",
    "requests>=2.31",
    "asyncssh>=2.14",
    "anthropic>=0.28",
    "openai>=1.30",
    "discord.py>=2.3",
    "python-telegram-bot>=20.7",
    "pyyaml>=6.0",
    "pydantic>=2.6",
    "rich>=13.7",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
janitor = "janitor.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
